+++ MenuSceneScripts/GameData.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public static class GameData
 {
-    public static int difficulty;
+    public static float difficulty;
 }
